+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- Improve primary domain parsing.
+- Refactor on-page email generation icon injection code.
+
 ## [0.0.1] - 2023-11-11
 
 - Release of something awesome.